+++ releasenotes.md
@@ -139,6 +139,7 @@
 - Packed structs sometimes not lowered as such.
 - Crash when creating `$Type*` where `$Type` is an optional type #2848
 - Crashes when using `io::EOF~!` in various unhandled places. #2848
+- Crash when trying to create a const zero untyped list #2847
 
 ### Stdlib changes
 - Add `ThreadPool` join function to wait for all threads to finish in the pool without destroying the threads.